Moved AnalyzerSrc class to python/analyzer
