+++ python/analyzer/__init__.py
@@ -1,6 +1,7 @@
 from .target_analyzer import AnalyzerCtlAddressMap,  AnalyzerCfgAddressMap, TbApbAddressMap
 from .target_analyzer_filter import Filter, FilterAcceptAll, FilterChanging
 from .target_analyzer_trigger import TriggerSimple
+from .analyzer_src import AnalyzerSrc
 
 from .analyzer import t_analyzer_data4, t_analyzer_filter_cfg, t_access_combs, t_analyzer_trace_cfg_fifo, t_address_op, t_alu_op, t_access_resp
 
@@ -9,4 +10,5 @@ __all__ += [AnalyzerCtlAddressMap, AnalyzerCfgAddressMap, TbApbAddressMap]
 __all__ += [t_analyzer_data4, t_analyzer_filter_cfg, t_access_combs, t_analyzer_trace_cfg_fifo, t_address_op, t_alu_op, t_access_resp]
 __all__ += [Filter, FilterAcceptAll, FilterChanging]
 __all__ += [TriggerSimple]
+__all__ += [AnalyzerSrc]
