+++ test/python/test_apb_analyzer_src.py
@@ -21,54 +21,13 @@ from regress.apb.bfm     import ApbMaster
 from queue import Queue
 from regress.utils import t_dprintf_req_4, t_dprintf_byte, Dprintf, t_dbg_master_request, t_dbg_master_response, DprintfBus, SramAccessBus, SramAccessRead, SramAccessWrite, DbgMaster, DbgMasterMuxScript, DbgMasterSramScript, DbgMasterFifoScript, FifoStatus, t_sram_access_req, t_sram_access_resp
 from regress.analyzer import TbApbAddressMap, Filter, t_analyzer_data4, FilterAcceptAll, FilterChanging
+from regress.analyzer import AnalyzerSrc
 
 from cdl.utils   import csr
 from cdl.sim     import ThExecFile, LogEventParser
 from cdl.sim     import HardwareThDut
 from cdl.sim     import TestCase
 from typing import Optional
-
-#c AnalyzerSrc
-class AnalyzerSrc:
-    data_cfg = [1,1,1,1]
-    tgt_mode = 1
-    def __init__(self, data=None, data_cfg=None):
-        self.data = [0,0,0,0]
-        if data is not None:
-            self.data = data[:]
-            pass
-        if data_cfg is not None:
-            self.data_cfg = data_cfg
-            pass
-        pass
-    #f apb_writes
-    def apb_writes(self, map):
-        cfg = self.tgt_mode
-        for i in range(4):
-            cfg = cfg | (self.data_cfg[i] << (16+i*4))
-            pass
-        writes = []
-        writes.append( (map.cfg, 0) )
-        writes.append( (map.data0, self.data[0] ) )
-        writes.append( (map.data1, self.data[1] ) )
-        writes.append( (map.data2, self.data[2] ) )
-        writes.append( (map.data3, self.data[3] ) )
-        # Write the cfg last so that the data does not start changing
-        writes.append((map.cfg, cfg))
-        return writes
-
-    def next_data(cfg, d):
-        if cfg == 1:
-            return d+1
-        return d
-    def next_valid(self):
-        while True:
-            for i in range(4):
-                self.data[i] = AnalyzerSrc.next_data(self.data_cfg[i], self.data[i])
-                pass
-            break
-        return tuple(self.data)
-    pass
 
 #c ApbAnalyzerTest_Base
 class ApbAnalyzerTest_Base(ThExecFile):
